"Refactor code to use Url_parameters instead of manual URL parsing and modify platform handling in App and AccountManager classes"
