+++ js/services/pageService.js
@@ -1,11 +1,7 @@
+import { Url_parameters } from './parameters.js';
+
 export function showPage(pageId) {
-    let url_string = window.location.href
-    let questionMarkCount = 0;
-    let modified_url = url_string.replace(/\?/g, function(match) {
-        questionMarkCount++;
-        return questionMarkCount === 2 ? '&' : match;
-    });
-    const url = new URL(modified_url);
+    const url = Url_parameters()
     const params = new URLSearchParams(url.search);
     const platform = params.get('platform');
